requirements.txt
python-dotenv
langchain-core
langgraph
langchain-openai
langchain-anthropic
langchain-chroma
chromadb
langchain-tavily
pinecone
langchain-pinecone
langgraph-checkpoint-postgres
pydantic
ipython
fastapi
uvicorn[standard]
psycopg[binary]